backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
google-generativeai==0.3.2
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
fpdf==1.7.2
